test: consolidate http2 tests in one file
PR-URL: https://github.com/nodejs/node/pull/15624
Refs: https://github.com/nodejs/node/issues/14985
